refactor: restructure RubyToBlocksConverter and implement local variable support
- Split RubyToBlocksConverter into specialized utility modules (block, node, variable, context).
- Restructured AST handlers into a modular directory (core, expressions, control-flow, assignments).
- Introduced ScopeManager for deterministic scope tracking during AST traversal.
- Implemented pseudo-local variable support using indexed instance variables and block metadata
comments.
- Updated Ruby generator to support local variable round-trip conversion.
- Ensured procedure arguments are handled consistently with Scratch VM.
