+++ packages/scratch-gui/src/lib/ruby-to-blocks-converter/ast-handlers/index.js
@@ -1,0 +1,18 @@
+import CoreHandlers from './core';
+import ExpressionHandlers from './expressions';
+import ControlFlowHandlers from './control-flow';
+import AssignmentHandlers from './assignments';
+
+/**
+ * Aggregated AST handlers for RubyToBlocksConverter.
+ * @mixes RubyToBlocksConverter
+ */
+const AstHandlers = Object.assign(
+    {},
+    CoreHandlers,
+    ExpressionHandlers,
+    ControlFlowHandlers,
+    AssignmentHandlers
+);
+
+export default AstHandlers;
